+++ apps/theme/components/Previews/Components/Components.tsx
@@ -35,7 +35,6 @@ import { useState } from 'react';
 import classes from './Components.module.css';
 
 export const Components = () => {
-  const [radioValue, setRadioValue] = useState('vanilje');
   const [currentPage, setCurrentPage] = useState(1);
   const pagination = usePagination({
     currentPage,
@@ -47,7 +46,7 @@ export const Components = () => {
   return (
     <div className={classes.components}>
       <div className={cl(classes.card, classes.checkbox)}>
-        <Checkbox.Group error='' legend='Handleliste' size='sm'>
+        <Fieldset legend='Handleliste' size='sm'>
           <Checkbox value='epost'>En kilo poteter</Checkbox>
           <Checkbox value='telefon'>To liter Farris</Checkbox>
           <Checkbox value='sms' defaultChecked>
@@ -61,7 +60,7 @@ export const Components = () => {
           </Checkbox>
           <Checkbox value='sms'>2kg smågodt</Checkbox>
           <Checkbox value='sms'>10 poser med Smash</Checkbox>
-        </Checkbox.Group>
+        </Fieldset>
       </div>
       <div className={cl(classes.card, classes.user)}>
         <Heading className={cl(classes.cardTitle, classes.userTitle)} size='xs'>
@@ -205,19 +204,18 @@ export const Components = () => {
         </div>
       </div>
       <div className={cl(classes.card, classes.radio)}>
-        <Radio.Group
-          error=''
+        <Fieldset
           legend='Hvilken iskremsmak er best?'
           description='Velg din favorittsmak'
           size='sm'
-          value={radioValue}
-          onChange={(e: string) => setRadioValue(e)}
         >
           <Radio value='vanilje'>Vanilje</Radio>
-          <Radio value='jordbær'>Jordbær</Radio>
+          <Radio value='jordbær' defaultChecked>
+            Jordbær
+          </Radio>
           <Radio value='sjokolade'>Sjokolade</Radio>
           <Radio value='spiser-ikke-is'>Jeg spiser ikke iskrem</Radio>
-        </Radio.Group>
+        </Fieldset>
       </div>
       <div className={cl(classes.card, classes.tag)}>
         <Heading size='xs' className={classes.tagHeading}>
