+++ packages/react/src/components/form/Checkbox/Checkbox.mdx
@@ -1,15 +1,14 @@
 import { Meta, Canvas, Controls, Primary } from '@storybook/blocks';
 import { Do, Dont, Stack, CssVariables } from '@doc-components';
-import css from '@digdir/designsystemet-css/checkbox.css?inline';
+import css from '@digdir/designsystemet-css/input.css?inline';
 import * as TableStories from '../../Table/Table.stories';
 import * as CheckboxStories from './Checkbox.stories';
-import * as CheckboxGroupStories from './CheckboxGroup.stories';
 
 <Meta of={CheckboxStories} />
 
 # Checkbox
 
-Vi bruker `Checkbox` (avmerkingsboks) for å gi brukerne valg, der de kan velge ett eller flere alternativer. Brukerne kan både velge og oppheve valgene de gjør. Bruk [`Checkbox.Group`](#checkboxgroup) til å gruppere flere valg.
+Vi bruker `Checkbox` (avmerkingsboks) for å gi brukerne valg, der de kan velge ett eller flere alternativer. Brukerne kan både velge og oppheve valgene de gjør. Bruk [`Fieldset`](#fieldset) til å gruppere flere valg.
 
 <Primary />
 <Controls />
@@ -19,12 +18,12 @@ Vi bruker `Checkbox` (avmerkingsboks) for å gi brukerne valg, der de kan velge 
 ```tsx
 import { Checkbox } from '@digdir/designsystemet-react';
 
-<Checkbox.Group legend='What would your D&D party like to eat?'>
+<Fieldset legend='What would your D&D party like to eat?'>
   <Checkbox value='elfsalad'>Wood Elf forest salad</Checkbox>
   <Checkbox value='firecakes'>Fire breath rice cakes</Checkbox>
   <Checkbox value='dumplings'>Chicken-something dumplings</Checkbox>
   <Checkbox value='ribs'>Sembian honey-glazed rothè ribs</Checkbox>
-</Checkbox.Group>;
+</Fieldset>;
 ```
 
 ## Eksempler
@@ -35,7 +34,7 @@ import { Checkbox } from '@digdir/designsystemet-react';
 
 Hvis brukeren skal bekrefte noe, men ikke velge noe, kan en `checkbox` stå alene. For eksempel når brukeren skal samtykke til vilkår.
 
-<Canvas of={CheckboxGroupStories.OneOption} />
+<Canvas of={CheckboxStories.OneOption} />
 
 ### Ubestemt
 
@@ -43,26 +42,26 @@ Bruk `indeterminate` til å signalisere at en eller flere sjekkbokser er markert
 
 <Canvas of={TableStories.WithFormElements} />
 
-### Checkbox.Group
+### Fieldset
 
-Bruk `Checkbox.Group` for gruppering av flere valg.
+Bruk `Fieldset` for gruppering av flere valg.
 
-<Canvas of={CheckboxGroupStories.Preview} />
-<Controls of={CheckboxGroupStories.Preview} />
+<Canvas of={CheckboxStories.Group} />
+<Controls of={CheckboxStories.Group} />
 
 #### Feilmelding
 
-Bruk `error` på `Checkbox.Group` for å vise feilmelding.
+Bruk `error` på `Fieldset` for å vise feilmelding.
 
-Her må du bruke `Checkbox.Group` for å gjøre innholdet tilgjengelig og aktivere riktig stil.
+Her må du bruke `Fieldset` for å gjøre innholdet tilgjengelig og aktivere riktig stil.
 
-<Canvas of={CheckboxGroupStories.WithError} />
+<Canvas of={CheckboxStories.WithError} />
 
 #### Kontrollert
 
-Bruk `value` på `Checkbox.Group` for å kontrollere verdiene selv.
+Bruk `value` på `Checkbox` for å kontrollere verdiene selv.
 
-<Canvas of={CheckboxGroupStories.Controlled} />
+<Canvas of={CheckboxStories.Controlled} />
 
 #### Readonly
 
@@ -70,13 +69,13 @@ Felter med `readonly`-attributtet er med i tabrekkefølgen. Brukerne kan kopiere
 
 `Readonly`-felter kan være forvirrende for noen brukere. Ikke alle vil skjønne hvorfor de ikke får til å endre innholdet i feltet. Vi anbefaler derfor å unngå readonly så langt det lar seg gjøre.
 
-<Canvas of={CheckboxGroupStories.ReadOnly} />
+<Canvas of={CheckboxStories.ReadOnly} />
 
 #### Disabled
 
 Unngå deaktiverte tilstander om du kan. De har lav fargekontrast som er problematisk for noen brukere. `Disabled` kan ikke møte kontrastkravene, for da kan brukeren tro at elementet er aktivt, prøve å trykke på det, og ikke skjønne hvorfor det ikke går.
 
-<Canvas of={CheckboxGroupStories.Disabled} />
+<Canvas of={CheckboxStories.Disabled} />
 
 <br />
 
@@ -139,7 +138,7 @@ Brukeren skal kunne lese hvert svaralternativ som en logisk respons på forklari
 
 Eksempel på riktig logisk sammenheng:
 
-<Canvas of={CheckboxGroupStories.ContentEx1} />
+<Canvas of={CheckboxStories.ContentEx1} />
 
 ### Gi svaralternativene samme språklige form
 
@@ -147,11 +146,11 @@ Det betyr at de kan ha enten fullstendige setninger eller ett ord, ikke begge de
 
 **Enten:**
 
-<Canvas of={CheckboxGroupStories.ContentEx2} />
+<Canvas of={CheckboxStories.ContentEx2} />
 
 **Eller:**
 
-<Canvas of={CheckboxGroupStories.ContentEx3} />
+<Canvas of={CheckboxStories.ContentEx3} />
 
 <br />
 
@@ -163,7 +162,7 @@ Det skal være mulig å flytte fokus til `Checkbox`, ved å trykke `Tab` på tas
 
 `Checkbox` skal alltid ha en `label`. Bruker du `label` fra en annen plass, husk da å bruke `aria-label` eller `aria-labelledby` (Se kode under).
 
-<Canvas of={CheckboxStories.Single} />
+<Canvas of={CheckboxStories.AriaLabel} />
 
 ## CSS Variabler
