+++ packages/react/src/components/form/Fieldset/Fieldset.tsx
@@ -18,9 +18,6 @@ export type FieldsetProps = {
   error?: ReactNode;
   /** The legend of the fieldset. */
   legend: ReactNode;
-  /** Toggle `readOnly` on fieldset context.
-   * @note This does not prevent fieldset values from being submited */
-  readOnly?: boolean;
   /** Visually hide `legend` and `description` (still available for screen readers)  */
   hideLegend?: boolean;
 } & Pick<FormFieldProps, 'size'> &
@@ -48,7 +45,6 @@ export const Fieldset = forwardRef<HTMLFieldSetElement, FieldsetProps>(
           disabled: props?.disabled,
           error: error ?? fieldset?.error,
           errorId: hasError ? errorId : undefined,
-          readOnly, // TODO should this be a prop on Fieldset?
           size,
         }}
       >
