+++ packages/react/src/components/form/Fieldset/FieldsetContext.ts
@@ -4,7 +4,7 @@ import type { FormFieldProps } from '../useFormField';
 
 export type FieldsetContextType = Pick<
   FormFieldProps,
-  'error' | 'errorId' | 'disabled' | 'readOnly' | 'size'
+  'error' | 'errorId' | 'disabled' | 'size'
 >;
 
 export const FieldsetContext = createContext<FieldsetContextType | null>(null);
