+++ packages/react/src/components/form/Input/Input.tsx
@@ -33,7 +33,15 @@ export type InputProps = {
  * ```
  */
 export const Input = forwardRef<HTMLInputElement, InputProps>(function Input(
-  { type = 'text', size = 'md', htmlSize, className, onClick, ...rest },
+  {
+    type = 'text',
+    size = 'md',
+    htmlSize,
+    className,
+    onChange,
+    onClick,
+    ...rest
+  },
   ref,
 ) {
   return (
@@ -43,6 +51,9 @@ export const Input = forwardRef<HTMLInputElement, InputProps>(function Input(
       ref={ref}
       size={htmlSize}
       type={type}
+      onChange={(event) => {
+        if (!rest.readOnly) onChange?.(event); // Make readonly work for checkbox / radio / switch
+      }}
       onClick={(event) => {
         if (rest.readOnly) event.preventDefault(); // Make readonly work for checkbox / radio / switch
         onClick?.(event);
