+++ packages/react/stories/showcase.stories.tsx
@@ -9,6 +9,7 @@ import {
   Checkbox,
   Combobox,
   Divider,
+  Fieldset,
   Heading,
   Link,
   Pagination,
@@ -54,7 +55,7 @@ export const Showcase: StoryFn = () => {
   return (
     <div className={classes.components}>
       <div className={cl(classes.card, classes.checkbox)}>
-        <Checkbox.Group error='' legend='Handleliste' size='sm'>
+        <Fieldset legend='Handleliste' size='sm'>
           <Checkbox value='epost'>En kilo poteter</Checkbox>
           <Checkbox value='telefon'>To liter Farris</Checkbox>
           <Checkbox value='sms' defaultChecked>
@@ -69,7 +70,7 @@ export const Showcase: StoryFn = () => {
           <Divider />
           <Checkbox value='sms'>2kg smågodt</Checkbox>
           <Checkbox value='sms'>10 poser med Smash</Checkbox>
-        </Checkbox.Group>
+        </Fieldset>
       </div>
       <div className={cl(classes.card, classes.user)}>
         <Heading className={cl(classes.cardTitle, classes.userTitle)} size='xs'>
@@ -214,18 +215,14 @@ export const Showcase: StoryFn = () => {
         </div>
       </div>
       <div className={cl(classes.card, classes.radio)}>
-        <Radio.Group
-          error=''
-          legend='Hvilken iskremsmak er best?'
-          size='sm'
-          value={radioValue}
-          onChange={(e: string) => setRadioValue(e)}
-        >
+        <Fieldset error='' legend='Hvilken iskremsmak er best?' size='sm'>
           <Radio value='vanilje'>Vanilje</Radio>
-          <Radio value='jordbær'>Jordbær</Radio>
+          <Radio value='jordbær' defaultChecked>
+            Jordbær
+          </Radio>
           <Radio value='sjokolade'>Sjokolade</Radio>
           <Radio value='spiser-ikke-is'>Jeg spiser ikke iskrem</Radio>
-        </Radio.Group>
+        </Fieldset>
       </div>
       <div className={cl(classes.card, classes.tag)}>
         <Heading size='xs'>Emner</Heading>
